kernel: x86_64: ps2hid: reverse IRQ ack and data read order
Hyper-V only seems to assert the mouse IRQ when it inserts a byte into
the emulated PS/2 data bus, and not while there is data available, so
if we acknowledge when there's data but we don't read it, then we will
never be re-interrupted to read it.

Resolve this by swapping the order; acknowledge the IRQ, then read a
byte. This order lets us correctly get interrupts for subsequent bytes.