kernel: x86_64: get TSC frequency from Hyper-V if available
Hyper-V's implementation of the PIT does not seem to completely support
channel 2; experimentation indicates that the counter is there and does
count down, but the output state (bit 5 in ioport 0x61) is fixed. There
appears to be no support in Hyper-V for a PC speaker, and this may be a
consequence of that design.

Hyper-V does expose an interface to guests to be able to query the TSC
frequency instead of guests needing to time it for themselves. If that
functionality is present, use it instead.